requirements.txt
playwright>=1.58.0
Flask>=3.1.0
requests>=2.32.0
python-dotenv>=1.0.0
